+++ lib/data/models/vit_table_column.dart
@@ -1,25 +1,24 @@
 import 'package:flutter/widgets.dart';
 
 class VitTableColumn {
-  final String title;
+  final Widget title;
 
   /// The width of the column.
   ///
-  /// If [expandable] is set to `true`, the this property is treated as minimum
-  /// width.
+  /// If [flex] is not null, the this property is treated as minimum width.
   final double width;
 
   /// Indicates that the column can be expanded if there are extra horizontal
   /// space.
   ///
   /// Has no effect if VitTable.enableHorizontalScroll is set to true.
-  final bool expandable;
+  final int? flex;
 
   final EdgeInsets cellsPadding;
 
   /// Used to hide columns.
   ///
-  /// If there is not enough space, the column with the greatest priority
+  /// If there is not enough space, the column with the greatest value
   /// will vanish first.
   ///
   /// Has no effect if VitTable.enableHorizontalScroll is set to true.
@@ -31,7 +30,7 @@ class VitTableColumn {
     required this.title,
     this.onSort,
     this.width = 150,
-    this.expandable = false,
+    this.flex,
     this.priority = 1,
     this.cellsPadding = const EdgeInsets.symmetric(horizontal: 12, vertical: 8),
   }) {