+++ lib/ui/theme/page_navigator_theme.dart
@@ -11,4 +11,20 @@ class PageNavigatorThemeData {
 
   final PageNavigatorStyle style;
   final PageNavigatorOptions options;
+
+  /// Merges the current PageNavigatorThemeData with another, prioritizing non-null values from the other instance.
+  PageNavigatorThemeData merge(PageNavigatorThemeData other) {
+    return PageNavigatorThemeData(
+      style: PageNavigatorStyle(
+        selectedColor: other.style.selectedColor ?? style.selectedColor,
+        color: other.style.color ?? style.color,
+        itemSize: other.style.itemSize ?? style.itemSize,
+      ),
+      options: PageNavigatorOptions(
+        jumpPageOffset: other.options.jumpPageOffset,
+        showEdgePages: other.options.showEdgePages,
+        showJumpPage: other.options.showJumpPage,
+      ),
+    );
+  }
 }